Rename other stack files and tests
